[lib redesign] stable version
