+++ app/src/main/java/com/pushpal/jetlime/ui/timelines/CustomizedVerticalTimeLine.kt
@@ -27,49 +27,75 @@ package com.pushpal.jetlime.ui.timelines
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.pushpal.jetlime.EventPointType
 import com.pushpal.jetlime.JetLimeColumn
 import com.pushpal.jetlime.JetLimeDefaults
+import com.pushpal.jetlime.JetLimeEvent
 import com.pushpal.jetlime.JetLimeEventDefaults
+import com.pushpal.jetlime.sample.R
 import com.pushpal.jetlime.ui.data.getCharacters
 import com.pushpal.jetlime.ui.timelines.event.VerticalEventContent
 
 @ExperimentalAnimationApi
 @Composable
-fun VerticalScrollableTimeLine(modifier: Modifier = Modifier) {
-  val listState = rememberLazyListState()
-  val items = remember { getCharacters() }
+fun CustomizedVerticalTimeLine(modifier: Modifier = Modifier) {
+  val items = remember { getCharacters().subList(0, 5) }
 
   Surface(
     modifier = modifier.fillMaxSize(),
   ) {
     JetLimeColumn(
-      modifier = Modifier.padding(32.dp),
-      listState = listState,
+      modifier = Modifier.padding(16.dp),
+      items = items,
+      keyExtractor = { item -> item.id },
       style = JetLimeDefaults.columnStyle(
+        contentDistance = 32.dp,
+        itemSpacing = 16.dp,
+        lineThickness = 2.dp,
         lineBrush = JetLimeDefaults.lineGradientBrush(),
+        pointStartFactor = 3f,
       ),
-    ) {
-      items.forEach { item ->
-        JetLimeEvent(
-          style = JetLimeEventDefaults.eventStyle(),
-        ) {
-          VerticalEventContent(item = item)
-        }
+    ) { index, item, position ->
+      JetLimeEvent(
+        style = JetLimeEventDefaults.eventStyle(
+          position = position,
+          pointColor = Color(0xFFFF9800),
+          pointFillColor = Color(0xFF006B04),
+          pointRadius = 10.dp,
+          pointAnimation = if (index == 2) JetLimeEventDefaults.pointAnimation() else null,
+          pointType = when (index) {
+            1 -> EventPointType.filled(0.8f)
+            4 -> EventPointType.custom(
+              icon = painterResource(id = R.drawable.icon_check),
+            )
+
+            else -> EventPointType.Default
+          },
+          pointStrokeWidth = when (index) {
+            4 -> 0.dp
+            else -> 2.dp
+          },
+          pointStrokeColor = Color(0xFFFFFFFF),
+        ),
+      ) {
+        VerticalEventContent(item = item)
       }
     }
   }
 }
 
 @ExperimentalAnimationApi
-@Preview("Preview VerticalScrollableTimeLine")
+@Preview("Preview CustomizedVerticalTimeLine")
 @Composable
-private fun PreviewSimpleVerticalScrollableTimeLine() {
-  VerticalScrollableTimeLine()
+private fun PreviewCustomizedVerticalTimeLine() {
+  CustomizedVerticalTimeLine()
 }
