+++ jetlime/src/main/java/com/pushpal/jetlime/JetLimeList.kt
@@ -30,12 +30,12 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.compositionLocalOf
-import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
@@ -54,16 +54,16 @@ import com.pushpal.jetlime.Arrangement.VERTICAL
  * @param content A composable lambda defining the content of the column.
  */
 @Composable
-fun JetLimeColumn(
+fun <T> JetLimeColumn(
   modifier: Modifier = Modifier,
   style: JetLimeStyle = JetLimeDefaults.columnStyle(),
   listState: LazyListState = rememberLazyListState(),
   contentPadding: PaddingValues = PaddingValues(0.dp),
-  content: @Composable JetLimeListScope.() -> Unit,
+  items: List<T>,
+  keyExtractor: (T) -> Any = {},
+  itemContent: @Composable (index: Int, T, EventPosition) -> Unit,
 ) {
   CompositionLocalProvider(LocalJetLimeStyle provides style.alignment(VERTICAL)) {
-    val items = remember { mutableStateListOf<@Composable (EventPosition) -> Unit>() }
-    JetLimeListScope(items).content()
     LazyColumn(
       modifier = modifier,
       state = listState,
@@ -74,8 +74,9 @@ fun JetLimeColumn(
       userScrollEnabled = true,
       contentPadding = contentPadding,
     ) {
-      items(items.size) { index ->
-        items[index](EventPosition.dynamic(index, items.size))
+      itemsIndexed(items = items, key = { _, item -> keyExtractor(item) }) { index, item ->
+        val eventPosition = EventPosition.dynamic(index, items.size)
+        itemContent(index, item, eventPosition)
       }
     }
   }
@@ -93,16 +94,16 @@ fun JetLimeColumn(
  * @param content A composable lambda defining the content of the row.
  */
 @Composable
-fun JetLimeRow(
+fun <T> JetLimeRow(
   modifier: Modifier = Modifier,
   style: JetLimeStyle = JetLimeDefaults.rowStyle(),
   listState: LazyListState = rememberLazyListState(),
   contentPadding: PaddingValues = PaddingValues(0.dp),
-  content: @Composable JetLimeListScope.() -> Unit,
+  items: List<T>,
+  keyExtractor: (T) -> Any = {},
+  itemContent: @Composable (index: Int, T, EventPosition) -> Unit,
 ) {
   CompositionLocalProvider(LocalJetLimeStyle provides style.alignment(HORIZONTAL)) {
-    val items = remember { mutableStateListOf<@Composable (EventPosition) -> Unit>() }
-    JetLimeListScope(items).content()
     LazyRow(
       modifier = modifier,
       state = listState,
@@ -113,8 +114,9 @@ fun JetLimeRow(
       userScrollEnabled = true,
       contentPadding = contentPadding,
     ) {
-      items(items.size) { index ->
-        items[index](EventPosition.dynamic(index, items.size))
+      itemsIndexed(items = items, key = { _, item -> keyExtractor(item) }) { index, item ->
+        val eventPosition = EventPosition.dynamic(index, items.size)
+        itemContent(index, item, eventPosition)
       }
     }
   }
